Huvudmeny samt delvis gjort ljudfunktion då ljudet stör plusknappens egentliga funktion, ljudfunktion kommenterad i calculationsact rätt långt upp på onCreate metoden
