+++ app/src/main/AndroidManifest.xml
@@ -8,12 +8,13 @@
         android:label="   Pyramid Test - V02max"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".UserProfileActivity" android:windowSoftInputMode="stateHidden" android:screenOrientation="portrait">
+        <activity android:name=".MenyActivity" android:windowSoftInputMode="stateHidden" android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".UserProfileActivity" android:windowSoftInputMode="stateHidden" android:screenOrientation="portrait"/>
         <activity android:name=".CalculationActivity" android:windowSoftInputMode="stateHidden" android:screenOrientation="portrait"/>
         <activity android:name=".PresentingActivity" android:label="  Home" android:windowSoftInputMode="stateHidden" android:screenOrientation="portrait"/>
         <activity android:name=".DataListActivity"/>
