+++ app/src/main/java/com/example/sam/vo2max/PresentingActivity.java
@@ -79,7 +79,7 @@ public class PresentingActivity extends AppCompatActivity {
     }
 
     public boolean onOptionsItemSelected(MenuItem item){
-        Intent myIntent = new Intent(getApplicationContext(), UserProfileActivity.class);
+        Intent myIntent = new Intent(getApplicationContext(), MenyActivity.class);
         startActivityForResult(myIntent, 0);
         return true;
