+++ app/src/main/res/layout/activity_main.xml
@@ -56,83 +56,53 @@
             android:layout_below="@+id/AgeTextView"
             android:layout_alignParentStart="true" />
 
-        <Button
-            android:id="@+id/Statistic_Btn"
-            android:layout_width="135dp"
-            android:layout_height="wrap_content"
-            android:layout_alignParentStart="true"
-            android:layout_below="@+id/NewMeasurementButtonID"
-            android:layout_marginTop="16dp"
-            android:background="@drawable/my_button_beige"
-            android:drawableLeft="@drawable/ic_listbutton_2_4"
-            android:onClick="viewContact"
-            android:paddingEnd="3dp"
-            android:paddingLeft="3dp"
-            android:text="@string/database_btn"
-            android:textColor="@color/black" />
-
         <EditText
             android:id="@+id/WeightEditTextID"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_alignBaseline="@+id/WeightTextViewID"
-            android:layout_alignBottom="@+id/WeightTextViewID"
-            android:layout_toEndOf="@+id/Statistic_Btn"
             android:ems="5"
             android:hint="@string/kg"
             android:inputType="number"
-            android:textAlignment="center" />
+            android:textAlignment="center"
+            android:layout_alignBaseline="@+id/WeightTextViewID"
+            android:layout_alignBottom="@+id/WeightTextViewID"
+            android:layout_alignEnd="@+id/NewMeasurementButtonID" />
 
         <EditText
             android:id="@+id/AgeEditTextID"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_above="@+id/WeightTextViewID"
-            android:layout_alignStart="@+id/WeightEditTextID"
             android:ems="5"
             android:hint="@string/user_age_hint"
             android:inputType="number"
-            android:textAlignment="center" />
+            android:textAlignment="center"
+            android:layout_above="@+id/WeightEditTextID"
+            android:layout_alignStart="@+id/WeightEditTextID" />
 
         <EditText
             android:id="@+id/NameEditTextID"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_alignStart="@+id/AgeEditTextID"
-            android:layout_below="@+id/OtherTestsCheckBoxID"
             android:ems="6"
             android:hint="@string/user_id_hint"
             android:inputType="none|textPersonName"
             android:selectAllOnFocus="true"
-            android:textAlignment="center" />
+            android:textAlignment="center"
+            android:layout_below="@+id/OtherTestsCheckBoxID"
+            android:layout_alignStart="@+id/AgeEditTextID" />
 
         <Button
             android:id="@+id/NewMeasurementButtonID"
-            android:layout_width="wrap_content"
+            android:layout_width="170dp"
             android:layout_height="wrap_content"
-            android:layout_alignEnd="@+id/WeightEditTextID"
-            android:layout_below="@+id/WeightEditTextID"
-            android:layout_marginTop="58dp"
-            android:background="@drawable/my_button_blue"
+            android:layout_marginTop="64dp"
+            android:background="@drawable/my_button_green"
             android:drawableStart="@drawable/ic_mesurmentbutton_2_4_2"
             android:paddingEnd="3dp"
             android:paddingStart="3dp"
-            android:text="@string/NewMeasurement" />
-
-        <Button
-            android:id="@+id/button"
-            android:layout_width="135dp"
-            android:layout_height="wrap_content"
-            android:layout_alignBaseline="@+id/Statistic_Btn"
-            android:layout_alignBottom="@+id/Statistic_Btn"
-            android:layout_alignParentEnd="true"
-            android:background="@drawable/my_button_beige"
-            android:drawableStart="@drawable/ic_searchbutton_2_4"
-            android:onClick="search_contact"
-            android:paddingEnd="3dp"
-            android:paddingStart="3dp"
-            android:text="@string/btnSearch"
-            android:textColor="@color/black" />
+            android:text="@string/NewMeasurement"
+            android:layout_below="@+id/WeightEditTextID"
+            android:layout_centerHorizontal="true" />
 
         <CheckBox
             android:text="@string/Pretest1"
