+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">Pyramid Test - V02max</string>
     <string name="greeting">Welcome to the swedish school of sport and health sciences (GIH) VO2max mobile application!</string>
-    <string name="NewMeasurement">New measurement</string>
+    <string name="NewMeasurement">Start test</string>
     <string name="agetextview">Enter Age:</string>
     <string name="Pretest1">Pretest 1</string>
     <string name="othertest">Other test?</string>
@@ -75,6 +75,10 @@
     <string name="user_id_hint">User ID</string>
     <string name="date">Date</string>
     <string name="date_2">Date</string>
+    <string name="StartNewMeasurementButton">Start a new measurement</string>
+    <string name="DatabaseButton">Database</string>
+    <string name="SearchContactButton">Search Contact</string>
+    <string name="InstructionsButtonID">Instructions</string>
 
 
 </resources>
